Se creo la clase Asigantura, heredando de la clase Alumno. Con sus respectivos atributos y metodos

+++ Examen/IAsignatura.cs
@@ -5,7 +5,7 @@ namespace Examen
     interface IAsignatura 
     {
         public double CalcularNotaFinal();
-        public double CalcularNotaFinal(double nota1, double nota2, double nota3);
+        public double CalcularNotaFinal(int nota1, int nota2, int nota3);
         public string MensajeNotaFinal(double notaFinal);
         public string Imprimir();
